Refactor hooks related to removing accounts
