+++ src/components/AccountDrawer/RenameRemoveMenuSwitch.tsx
@@ -4,28 +4,35 @@ import { useNavigate } from "react-router-dom";
 import { RenameAccountModal } from "./RenameAccountModal";
 import { Account } from "../../types/Account";
 import { useImplicitAccounts } from "../../utils/hooks/getAccountDataHooks";
-import { useAppDispatch } from "../../utils/redux/hooks";
-import { accountsSlice } from "../../utils/redux/slices/accountsSlice";
-import { remove as removeSecretKeyAccount } from "../../utils/redux/thunks/secretKeyAccount";
+import { useRemoveAccount } from "../../utils/hooks/removeAccountDataHooks";
 import { ConfirmationModal } from "../ConfirmationModal";
 import { DynamicModalContext } from "../DynamicModal";
 import { RenameRemoveMenu } from "../RenameRemoveMenu";
 
 export const RenameRemoveMenuSwitch: React.FC<{ account: Account }> = ({ account }) => {
   const { openWith, onClose: closeModal } = useContext(DynamicModalContext);
-  const dispatch = useAppDispatch();
   const navigate = useNavigate();
   const isLastImplicitAccount = useImplicitAccounts().length === 1;
+  const removeAccount = useRemoveAccount();
 
-  let removeAccountAction: any = undefined;
-
+  let onRemove = undefined;
   switch (account.type) {
     case "secret_key":
-      removeAccountAction = removeSecretKeyAccount;
-      break;
     case "ledger":
     case "social":
-      removeAccountAction = () => accountsSlice.actions.removeAccount(account);
+      onRemove = () =>
+        openWith(
+          <ConfirmationModal
+            buttonLabel={buttonLabel}
+            description={description}
+            onSubmit={() => {
+              removeAccount(account);
+              closeModal();
+              navigate("/");
+            }}
+            title={title}
+          />
+        );
       break;
     case "mnemonic":
     case "multisig":
@@ -42,22 +49,6 @@ export const RenameRemoveMenuSwitch: React.FC<{ account: Account }> = ({ account
       "Personal data (including saved contacts, password and accounts) won't be affected.";
     buttonLabel = "Remove & Off-board";
   }
-
-  const onRemove = !removeAccountAction
-    ? undefined
-    : () =>
-        openWith(
-          <ConfirmationModal
-            buttonLabel={buttonLabel}
-            description={description}
-            onSubmit={() => {
-              dispatch(removeAccountAction(account));
-              closeModal();
-              navigate("/");
-            }}
-            title={title}
-          />
-        );
 
   return (
     <RenameRemoveMenu
